refactor(clis,tools,types): Move TransactionReceipt to types
