+++ src/ethereum_clis/tests/test_execution_specs.py
@@ -34,7 +34,13 @@ def monkeypatch_path_for_entry_points(monkeypatch):
     monkeypatch.setenv("PATH", f"{bin_dir}:{os.environ['PATH']}")
 
 
-@pytest.mark.parametrize("t8n", [ExecutionSpecsTransitionTool()])
+@pytest.fixture
+def t8n(request):
+    """Fixture for the `t8n` argument."""
+    return request.param()
+
+
+@pytest.mark.parametrize("t8n", [ExecutionSpecsTransitionTool], indirect=True)
 @pytest.mark.parametrize("fork", [London, Istanbul])
 @pytest.mark.parametrize(
     "alloc,base_fee,expected_hash",
@@ -150,7 +156,7 @@ def env(test_dir: str) -> Environment:
         return Environment.model_validate_json(f.read())
 
 
-@pytest.mark.parametrize("t8n", [ExecutionSpecsTransitionTool()])
+@pytest.mark.parametrize("t8n", [ExecutionSpecsTransitionTool], indirect=True)
 @pytest.mark.parametrize("test_dir", os.listdir(path=FIXTURES_ROOT))
 def test_evm_t8n(
     t8n: TransitionTool,
