+++ src/ethereum_test_tools/__init__.py
@@ -51,6 +51,7 @@ from ethereum_test_types import (
     Storage,
     TestParameterGroup,
     Transaction,
+    TransactionReceipt,
     Withdrawal,
     WithdrawalRequest,
     add_kzg_version,
@@ -144,6 +145,7 @@ __all__ = (
     "TestPrivateKey2",
     "Transaction",
     "TransactionException",
+    "TransactionReceipt",
     "TransactionTest",
     "TransactionTestFiller",
     "Withdrawal",
