+++ src/ethereum_test_types/__init__.py
@@ -22,6 +22,7 @@ from .types import (
     Storage,
     Transaction,
     TransactionDefaults,
+    TransactionReceipt,
     Withdrawal,
     WithdrawalRequest,
     keccak256,
@@ -49,6 +50,7 @@ __all__ = (
     "TestPrivateKey2",
     "Transaction",
     "TransactionDefaults",
+    "TransactionReceipt",
     "Withdrawal",
     "WithdrawalRequest",
     "ZeroPaddedHexNumber",
