+++ src/ethereum_test_types/helpers.py
@@ -78,7 +78,7 @@ def compute_eofcreate_address(
 
 def add_kzg_version(
     b_hashes: List[bytes | SupportsBytes | int | str], kzg_version: int
-) -> List[bytes]:
+) -> List[Hash]:
     """Add  Kzg Version to each blob hash."""
     kzg_version_hex = bytes([kzg_version])
     kzg_versioned_hashes = []
@@ -86,11 +86,11 @@ def add_kzg_version(
     for b_hash in b_hashes:
         b_hash = bytes(Hash(b_hash))
         if isinstance(b_hash, int) or isinstance(b_hash, str):
-            kzg_versioned_hashes.append(kzg_version_hex + b_hash[1:])
+            kzg_versioned_hashes.append(Hash(kzg_version_hex + b_hash[1:]))
         elif isinstance(b_hash, bytes) or isinstance(b_hash, SupportsBytes):
             if isinstance(b_hash, SupportsBytes):
                 b_hash = bytes(b_hash)
-            kzg_versioned_hashes.append(kzg_version_hex + b_hash[1:])
+            kzg_versioned_hashes.append(Hash(kzg_version_hex + b_hash[1:]))
         else:
             raise TypeError("Blob hash must be either an integer, string or bytes")
     return kzg_versioned_hashes
